+++ ipam.go
@@ -1,7 +1,5 @@
 package ipam
 
-import "fmt"
-
 // Ipamer can be used to do IPAM stuff.
 type Ipamer struct {
 	storage Storage
@@ -17,21 +15,4 @@ func New() *Ipamer {
 // The Storage interface must be implemented.
 func NewWithStorage(storage Storage) *Ipamer {
 	return &Ipamer{storage: storage}
-}
-
-// NotFoundError is raised if the given Prefix or Cidr was not found
-type NotFoundError struct {
-	msg string
-}
-
-func (n NotFoundError) Error() string {
-	return n.msg
-}
-
-func newNotFoundError(msg string) NotFoundError {
-	return NotFoundError{msg: msg}
-}
-
-func newNotFoundErrorf(msg string, a ...interface{}) NotFoundError {
-	return NotFoundError{msg: fmt.Sprintf(msg, a...)}
 }
